+++ ExtraNotification/ExtraNotification.cs
@@ -1,4 +1,5 @@
 ﻿using HarmonyLib;
+using UnityEngine;
 using VoxelTycoon;
 using VoxelTycoon.Modding;
 
@@ -6,7 +7,7 @@ namespace ExtraNotification
 {
     class ExtraNotification : Mod
     {
-        private static readonly Logger _logger = new Logger<ExtraNotification>();
+        private static readonly VoxelTycoon.Logger _logger = new Logger<ExtraNotification>();
         private const string PatchId = "org.hertsig.voxeltycoon.ExtraNotification";
 
         protected override void Initialize()
@@ -15,13 +16,19 @@ namespace ExtraNotification
             _logger.Log("ExtraNotification patch complete");
         }
 
-        protected override void OnGameStarted()
+        protected override void OnUpdate()
         {
-            EmptyUnloadPatch.GameStarted = true;
+            //_logger.Log($"{Time.time}; {Time.timeSinceLevelLoad}; {Time.frameCount}");
+            if (!EmptyUnloadPatch.GameStarted && Time.timeSinceLevelLoad > 10)
+            {
+                _logger.Log($"{EmptyUnloadPatch.GameStarted} to true");
+                EmptyUnloadPatch.GameStarted = true;
+            }
         }
 
         protected override void Deinitialize()
         {
+            _logger.Log($"{EmptyUnloadPatch.GameStarted} to false");
             EmptyUnloadPatch.GameStarted = false;
             new Harmony(PatchId).UnpatchAll();
             _logger.Log("ExtraNotification un-patch complete");