+++ ExtraNotification/changelog.md
@@ -1,12 +1,16 @@
-﻿# Changelog
+# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Changes
+
+### Fixed
+- Improved start-of-game check so the empty unload notification shouldn't trigger on game load any more.
+### Changed
 - Added custom icons to all notifications
+- Changed default value for Waiting at signal warning time from 2 to 3. Only affects new games.
 
 ## [0.1.0] - 2022-03-12
 First release